info card steps component - move to own file
