+++ packages/renderer/src/renderer/screens/Import/Ledger/Import/index.tsx
@@ -12,7 +12,6 @@ import { useAddresses } from '@app/contexts/import/Addresses';
 import { useConnections } from '@app/contexts/common/Connections';
 import { useImportHandler } from '@app/contexts/import/ImportHandler';
 import { useLedgerHardware } from '@ren/renderer/contexts/import/LedgerHardware';
-import { AnimatePresence, motion } from 'framer-motion';
 
 import { BarLoader } from 'react-spinners';
 import {
@@ -32,8 +31,6 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faCaretLeft,
   faCaretRight,
-  faCircleChevronLeft,
-  faCircleChevronRight,
   faCircleDot,
   faExclamationTriangle,
   faX,
@@ -42,12 +39,12 @@ import {
   CheckboxRoot,
   ConnectButton,
   InfoCard,
-  InfoCardStepsWrapper,
   LedgerAddressRow,
   SelectTrigger,
   SelectContent,
   AddressListFooter,
 } from './Wrappers';
+import { InfoCardSteps } from './InfoCardSteps';
 import { ContentWrapper } from '../../../Wrappers';
 import { determineStatusFromCodes } from './Utils';
 import { ItemsColumn } from '@app/screens/Home/Manage/Wrappers';
@@ -71,63 +68,6 @@ const SelectItem = forwardRef(function SelectItem(
     </Select.Item>
   );
 });
-
-const InfoCardSteps = ({
-  children,
-  style,
-  className,
-}: {
-  children: React.ReactNode;
-  style?: React.CSSProperties;
-  className?: string;
-}) => {
-  const [stepIndex, setStepIndex] = useState(0);
-  const totalSteps = 3;
-
-  const getSteps = (): React.ReactNode[] =>
-    (children as React.ReactNode[]).map((fragment, i) => (
-      <AnimatePresence key={`step-${i}`}>
-        <motion.span
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ duration: 0.75 }}
-        >
-          <FontAwesomeIcon icon={faCircleDot} transform={'shrink-3'} />
-          {fragment}
-        </motion.span>
-      </AnimatePresence>
-    ));
-
-  const clickChev = (dir: 'next' | 'prev') => {
-    setStepIndex((pv) => {
-      if (dir === 'next') {
-        return pv + 1 === totalSteps ? 0 : pv + 1;
-      } else {
-        return pv - 1 < 0 ? totalSteps - 1 : pv - 1;
-      }
-    });
-  };
-
-  return (
-    <InfoCardStepsWrapper className={className} style={style}>
-      {getSteps()[stepIndex]}
-      <div>
-        <button disabled={stepIndex === 0} onClick={() => clickChev('prev')}>
-          <FontAwesomeIcon className="chev" icon={faCircleChevronLeft} />
-        </button>
-        <span>
-          {stepIndex + 1} of {totalSteps}
-        </span>
-        <button
-          disabled={stepIndex === totalSteps - 1}
-          onClick={() => clickChev('next')}
-        >
-          <FontAwesomeIcon className="chev" icon={faCircleChevronRight} />
-        </button>
-      </div>
-    </InfoCardStepsWrapper>
-  );
-};
 
 export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
   const { darkMode } = useConnections();
